# client/src/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Assembled Intraday",
    "version": "0.1.0"
  },
  "paths": {
...
    "/notifications": {
      "get": {
        "tags": [
          "notifications"
        ],
        "summary": "List Notifications",
        "operationId": "list_notifications_notifications_get",
        "parameters": [
          {
            "name": "X-Username",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Username"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NotificationRead"
                  },
                  "title": "Response List Notifications Notifications Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "notifications"
        ],
        "summary": "Clear Inbox",
        "operationId": "clear_inbox_notifications_delete",
        "parameters": [
          {
            "name": "X-Username",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Username"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChannelType": {
        "type": "string",
        "enum": [
          "console",
          "inbox"
        ],
        "title": "ChannelType",
        "description": "Delivery channels for a notification."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "NotificationRead": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "rule_id": {
            "type": "string",
            "title": "Rule Id"
          },
          "recipient_id": {
            "type": "string",
            "title": "Recipient Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "body": {
            "type": "string",
            "title": "Body"
          },
          "severity": {
            "$ref": "#/components/schemas/Severity"
          },
          "entity_key": {
            "type": "string",
            "title": "Entity Key"
          },
          "triggering_event_id": {
            "type": "string",
            "title": "Triggering Event Id"
          },
          "ts": {
            "type": "string",
            "format": "date-time",
            "title": "Ts"
          },
          "delivered_channels": {
            "items": {
              "$ref": "#/components/schemas/ChannelType"
            },
            "type": "array",
            "title": "Delivered Channels"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "rule_id",
          "recipient_id",
          "title",
          "body",
          "severity",
          "entity_key",
          "triggering_event_id",
          "ts",
          "delivered_channels"
        ],
        "title": "NotificationRead",
        "description": "Persisted inbox notification."
      },
...
      "Severity": {
        "type": "string",
        "enum": [
          "info",
          "warning",
          "critical"
        ],
        "title": "Severity",
        "description": "Alert severity copied onto fired notifications."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}